requirements.txt
colorama==0.4.6
einops
GitPython==3.1.37
gym==0.17.3
h5py
huggingface-hub==0.17.3
hypothesis==6.87.1
imageio
imageio-ffmpeg
ipython
jupyter==1.0.0
matplotlib==3.7.2
mediapy
multiprocess==0.70.15
numpy==1.23.4
omegaconf
opencv-python==4.5.1.48
pandas
Pillow
pybullet==3.2.5
pyogrio==0.6.0
PyOpenGL==3.1.7
pyparsing==3.0.9
pyproj==3.6.1
pyproject_hooks==1.1.0
pyquaternion==0.9.9
pytest==7.4.2
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.1
sortedcontainers==2.4.0
termcolor==2.4.0
testbook==0.4.2
threadpoolctl==3.2.0
timm==0.9.7
tqdm==4.65.0
transforms3d==0.4.1
typed-argument-parser==1.10.0
wandb==0.14.2
